[FEATURE] Add default domain root pid config setting
This makes sure that the current indexed page/folder loads the typoscript from the default domain root pid, if the current page/folder is not in the rootline of the domain root pid.
Fixes e.g. correct realurl typolink generation, if the records are stored in page or folder  outside of the domain root rootline.

- The default domain root pid can be set in the extension configuration.
- useConfigurationFromClosestTemplate needs to be enabled also
- If default domain root pid is empty, the original processing from ext:solr takes place